run loop until out of work, add upper limit
The goal of running the loop until we're out of work is to emulate
exactly the same state that Weaviate would do if it ran the loop
forever.

The upper limit helps to prevent the script from copying 1TB of data
before it can even get started. This way, we can limit it to the first
100 or so files and it can already provide value. This makes the whole
operation less risky because it doesn't run four hours any time you
execute it.